+++ src/eduid_action/common/action_abc.py
@@ -31,6 +31,7 @@
 #
 from abc import ABCMeta, abstractmethod
 import gettext
+from flask import current_app
 
 
 class ActionError(Exception):
@@ -123,7 +124,6 @@ class ActionPlugin:
         :type app: flask.App
         '''
 
-    @abstractmethod
     def get_number_of_steps(self):
         '''
         The number of steps that the user has to take
@@ -134,8 +134,8 @@ class ActionPlugin:
         :returns: the number of steps
         :rtype: int
         '''
+        return self.steps
 
-    @abstractmethod
     def get_url_for_bundle(self, action):
         '''
         Return the url for the bundle that contains the front-end javascript
@@ -149,6 +149,15 @@ class ActionPlugin:
         :type action: dict
         :rtype: unicode
         '''
+        base = current_app.config.get('BUNDLES_URL')
+        bundle_name = '{}.js'
+        if current_app.config.get('DEBUG'):
+            bundle_name = '{}-bundle.dev.js'
+        url = '{}{}'.format(
+                base,
+                bundle_name.format(self.PACKAGE_NAME)
+                )
+        return url
 
     @abstractmethod
     def get_config_for_bundle(self, action):
